+++ app/src/main/java/com/websarva/wings/android/abiityofbaseball/listview/PlayerItemData.kt
@@ -1,4 +1,4 @@
-package com.websarva.wings.android.abiityofbaseball
+package com.websarva.wings.android.abiityofbaseball.listview
 
 data class PlayerItemData(
         val playerId: Int,
